+++ src/settings/StaffSlips/StaffSlipForm.test.js
@@ -9,62 +9,99 @@ import '../../../test/jest/__mock__';
 
 import { Field } from 'react-final-form';
 
-import { TemplateEditor } from '@folio/stripes-template-editor';
 import {
-  KeyValue,
   Pane,
   Paneset,
-  TextArea,
+  Accordion,
+  AccordionSet,
+  ExpandAllButton,
 } from '@folio/stripes/components';
 
 import StaffSlipForm from './StaffSlipForm';
-import TokensList from './TokensList';
-import getTokens from './tokens';
+import {
+  StaffSlipAboutSection,
+  StaffSlipTemplateContentSection,
+} from './components/EditSections';
 import {
   CancelButton,
   FooterPane,
+  Metadata,
 } from '../components';
+
+const mockGeneralStaffSlipDetailId = 'generalInformation';
+const mockTemplateContentId = 'templateContent';
 
 jest.mock('../components', () => ({
   CancelButton: jest.fn(() => null),
   FooterPane: jest.fn(() => null),
+  Metadata: jest.fn(() => null),
+}));
+jest.mock('./components/EditSections', () => ({
+  StaffSlipAboutSection: jest.fn(() => 'StaffSlipAboutSection'),
+  StaffSlipTemplateContentSection: jest.fn(() => 'StaffSlipTemplateContentSection'),
 }));
 Field.mockImplementation(jest.fn(() => null));
 
+const mockTestIds = {
+  expandAllButton: 'expandAllButton',
+  accordion: 'accordion',
+  accordionSet: 'accordionSet',
+};
+
+ExpandAllButton.mockImplementation(({ onToggle }) => (
+  // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions
+  <div
+    data-testid={mockTestIds.expandAllButton}
+    onClick={() => onToggle({
+      generalInformation: false,
+      templateContent: false,
+    })}
+  />
+));
+
 describe('StaffSlipForm', () => {
-  const orderOfFieldExecution = {
-    description: 1,
-    template: 2,
-  };
   const labelIds = {
-    name: 'ui-circulation.settings.staffSlips.name',
-    description: 'ui-circulation.settings.staffSlips.description',
-    display: 'ui-circulation.settings.staffSlips.display',
-    preview: 'ui-circulation.settings.staffSlips.form.previewLabel',
     new: 'ui-circulation.settings.staffSlips.new',
+    staffSlipsGeneralInformation: 'ui-circulation.settings.staffSlips.generalInformation',
+    staffSlipsTemplateContent: 'ui-circulation.settings.staffSlips.templateContent',
   };
   const mockedHandleSubmit = jest.fn();
   const mockedOnCancel = jest.fn();
+  const mockedHandleExpandAll = jest.fn();
+  const mockedHandleSectionToggle = jest.fn();
 
   afterEach(() => {
-    Field.mockClear();
     Pane.mockClear();
+    Accordion.mockClear();
+    AccordionSet.mockClear();
+    ExpandAllButton.mockClear();
+    Metadata.mockClear();
     mockedHandleSubmit.mockClear();
     mockedOnCancel.mockClear();
+    mockedHandleExpandAll.mockClear();
+    mockedHandleSectionToggle.mockClear();
+    StaffSlipAboutSection.mockClear();
+    StaffSlipTemplateContentSection.mockClear();
   });
 
   describe('with main props', () => {
     const mockedInitialValues = {
       id: 'testId',
       name: 'testName',
+      metadata: 'testMetadata',
     };
     const mockedStripes = {
       hasPerm: jest.fn(() => true),
+      connect: jest.fn(),
     };
     const mockedfooterPaneProps = {
       pristine: true,
       submitting: true,
       onCancel: mockedOnCancel,
+    };
+    const accordionDefaultStatus = {
+      generalInformation: true,
+      templateContent: true,
     };
 
     beforeEach(() => {
@@ -106,41 +143,103 @@ describe('StaffSlipForm', () => {
       expect(FooterPane).toHaveBeenLastCalledWith(mockedfooterPaneProps, {});
     });
 
-    it('should execute "KeyValue" with passed props', () => {
-      const expectedResult = {
-        label: labelIds.name,
-        value: mockedInitialValues.name,
-      };
-
-      expect(KeyValue).toHaveBeenLastCalledWith(expectedResult, {});
+    it('should render AccordionSet component', () => {
+      expect(AccordionSet).toHaveBeenCalled();
     });
 
-    it('should execute "Field" associated with description with passed props', () => {
-      const expectedResult = {
-        label: labelIds.description,
-        name: 'description',
-        id: 'input-staff-slip-description',
-        autoFocus: true,
-        component: TextArea,
-        fullWidth: true,
-        disabled: false,
-      };
-
-      expect(Field).toHaveBeenNthCalledWith(orderOfFieldExecution.description, expectedResult, {});
+    it('should render Accordion component 2 times', () => {
+      expect(Accordion).toHaveBeenCalledTimes(2);
     });
 
-    it('should execute "Field" associated with template with passed props', () => {
-      const expectedResult = {
-        label: labelIds.display,
-        component: TemplateEditor,
-        tokens: getTokens('en'),
-        name: 'template',
-        previewModalHeader: labelIds.preview,
-        tokensList: TokensList,
-        printable: true,
-      };
+    it('should render "General information" label', () => {
+      expect(screen.getByText(labelIds.staffSlipsGeneralInformation)).toBeDefined();
+    });
 
-      expect(Field).toHaveBeenNthCalledWith(orderOfFieldExecution.template, expectedResult, {});
+    it('should render "Template content" label', () => {
+      expect(screen.getByText(labelIds.staffSlipsTemplateContent)).toBeDefined();
+    });
+
+    it("should render 'General information' Accordion", () => {
+      expect(Accordion).toHaveBeenNthCalledWith(1, expect.objectContaining({
+        id: mockGeneralStaffSlipDetailId,
+        open: accordionDefaultStatus.generalInformation,
+      }), {});
+    });
+
+    it('should render "Template content" Accordion', () => {
+      expect(Accordion).toHaveBeenNthCalledWith(2, expect.objectContaining({
+        id: mockTemplateContentId,
+        open: accordionDefaultStatus.templateContent,
+      }), {});
+    });
+
+    it('should call Metadata component', () => {
+      expect(Metadata).toHaveBeenCalledWith(expect.objectContaining({
+        connect: mockedStripes.connect,
+        metadata: mockedInitialValues.metadata,
+      }), {});
+    });
+
+    describe('handleExpandAll method', () => {
+      it('should render components with default accordions statuses', () => {
+        expect(ExpandAllButton).toHaveBeenLastCalledWith(expect.objectContaining({
+          accordionStatus: accordionDefaultStatus,
+        }), {});
+
+        expect(Accordion).toHaveBeenLastCalledWith(
+          expect.objectContaining({
+            open: accordionDefaultStatus.generalInformation,
+          }), {}
+        );
+      });
+
+      it('should expand all accordions statuses', () => {
+        fireEvent.click(screen.getByTestId(mockTestIds.expandAllButton));
+
+        expect(ExpandAllButton).toHaveBeenLastCalledWith(expect.objectContaining({
+          accordionStatus: {
+            generalInformation: false,
+            templateContent: false,
+          },
+        }), {});
+      });
+    });
+
+    describe('handleSectionToggle method', () => {
+      it('should close accordion', () => {
+        Accordion.mockImplementationOnce(({ onToggle, children }) => (
+          // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions
+          <div
+            data-testid={mockTestIds.accordion}
+            onClick={() => onToggle({ id: mockGeneralStaffSlipDetailId })}
+          >
+            {children}
+          </div>
+        ));
+
+        render(
+          <StaffSlipForm
+            stripes={mockedStripes}
+            handleSubmit={mockedHandleSubmit}
+            initialValues={mockedInitialValues}
+            {...mockedfooterPaneProps}
+          />
+        );
+
+        expect(Accordion).toHaveBeenCalledWith(
+          expect.objectContaining({
+            open: accordionDefaultStatus.generalInformation,
+          }), {}
+        );
+
+        fireEvent.click(screen.getByTestId(mockTestIds.accordion));
+
+        expect(Accordion).toHaveBeenCalledWith(
+          expect.objectContaining({
+            open: !accordionDefaultStatus.generalInformation,
+          }), {}
+        );
+      });
     });
   });
 
@@ -179,14 +278,6 @@ describe('StaffSlipForm', () => {
       );
       expect(FooterPane).toHaveBeenLastCalledWith(
         mockedfooterPaneProps,
-        {},
-      );
-    });
-
-    it('should execute "Field" associated with description with passed props', () => {
-      expect(Field).toHaveBeenNthCalledWith(
-        orderOfFieldExecution.description,
-        expect.objectContaining({ disabled: true }),
         {},
       );
     });