+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## IN PROGRESS
 * Add metadata information to view of Staff Slips (Settings > Circulation > Staff Slips). Refs UICIRC-855.
+* Implement General information accordion. Refs UICIRC-909.
+* Implement Template content accordion. Refs UICIRC-910.
 * Add metadata info to view of Patron Notice Templates (Settings > Circulation > Patron Notice Templates). Refs UICIRC-856.
 
 ## [8.0.0](https://github.com/folio-org/ui-circulation/tree/v8.0.0) (2023-02-22)